feat(domain): introduce shared test_strategies module
Hoists `arb_risk_level`, `arb_verdict`, and `arb_analysis_result` from
`src/adapters/reporters/json.rs::proptests` (where they were
module-private) up to a crate-root `#[cfg(test)] pub(crate)` module.
Also widens the analysis-result vec bound from `0..10` to `0..50` so
property tests probe empty, small-N, and N>typical-limit cases (CQO-C7),
and adds `arb_verdict_with_nan_coverage` for NaN-aware filter/sort tests
that follow in `domain::view`.

Reporter proptests now `use crate::test_strategies::arb_analysis_result`.

Also includes:
- Picks up the v0.1.1 version-bump in `full_table_snapshot.snap` (was
  pre-broken on `main` after release 8e3f25e).
- Brings the spec-only feature files (`view.feature`, `cli_ergonomics.feature`,
  `lcov_parser.feature`) into the worktree alongside the matching tests.

Closes #78 (foundation; behavior-preserving refactor follows in subsequent
commits on this branch).
