+++ src/adapters/reporters/json.rs
@@ -323,101 +323,8 @@ mod tests {
 #[cfg(test)]
 mod proptests {
     use super::*;
-    use crate::domain::types::{
-        AnalysisSummary, CrapScore, FunctionIdentity, FunctionVerdict, RiskDistribution, RiskLevel,
-        ScoredFunction, SourceSpan,
-    };
+    use crate::test_strategies::arb_analysis_result;
     use proptest::prelude::*;
-
-    fn arb_risk_level() -> impl Strategy<Value = RiskLevel> {
-        prop_oneof![
-            Just(RiskLevel::Low),
-            Just(RiskLevel::Acceptable),
-            Just(RiskLevel::Moderate),
-            Just(RiskLevel::High),
-        ]
-    }
-
-    fn arb_verdict() -> impl Strategy<Value = FunctionVerdict> {
-        (
-            "[a-z_]{1,20}",
-            "src/[a-z/]{1,30}\\.rs",
-            1..100u32,
-            0.0..=100.0f64,
-            1.0..200.0f64,
-            arb_risk_level(),
-            1.0..100.0f64,
-        )
-            .prop_map(
-                |(name, file, complexity, coverage, crap_value, risk, threshold)| FunctionVerdict {
-                    scored: ScoredFunction {
-                        identity: FunctionIdentity {
-                            file_path: file,
-                            qualified_name: name,
-                            span: SourceSpan {
-                                start_line: 1,
-                                end_line: 10,
-                            },
-                        },
-                        complexity,
-                        complexity_metric: ComplexityMetric::Cognitive,
-                        coverage_percent: coverage,
-                        crap: CrapScore {
-                            value: crap_value,
-                            risk_level: risk,
-                        },
-                        contributors: vec![],
-                    },
-                    threshold,
-                    exceeds: crap_value > threshold,
-                },
-            )
-    }
-
-    /// Build an AnalysisResult with a hand-constructed summary.
-    /// Summary values are structurally valid but not semantically precise —
-    /// reporters only format what they receive, so accuracy doesn't matter.
-    fn arb_analysis_result() -> impl Strategy<Value = crate::domain::types::AnalysisResult> {
-        prop::collection::vec(arb_verdict(), 0..10).prop_map(|verdicts| {
-            let total = verdicts.len();
-            let exceeding = verdicts.iter().filter(|v| v.exceeds).count();
-            let passed = exceeding == 0;
-            let max_crap = verdicts
-                .iter()
-                .max_by(|a, b| {
-                    a.scored
-                        .crap
-                        .value
-                        .partial_cmp(&b.scored.crap.value)
-                        .unwrap_or(std::cmp::Ordering::Equal)
-                })
-                .map(|v| v.scored.crap);
-            let avg = if total > 0 {
-                verdicts.iter().map(|v| v.scored.crap.value).sum::<f64>() / total as f64
-            } else {
-                0.0
-            };
-            crate::domain::types::AnalysisResult {
-                functions: verdicts,
-                summary: AnalysisSummary {
-                    total_functions: total,
-                    total_files: total,
-                    exceeding_threshold: exceeding,
-                    average_crap: avg,
-                    median_crap: avg,
-                    max_crap,
-                    worst_function: None,
-                    distribution: RiskDistribution {
-                        low: 0,
-                        acceptable: 0,
-                        moderate: 0,
-                        high: 0,
-                    },
-                },
-                passed,
-            }
-        })
-    }
 
     fn arb_config() -> impl Strategy<Value = JsonConfig<'static>> {
         (1.0..100.0f64,).prop_map(|(threshold,)| JsonConfig {
