+++ src/lib.rs
@@ -2,3 +2,6 @@ pub mod adapters;
 pub mod core;
 pub mod domain;
 pub mod ports;
+
+#[cfg(test)]
+pub(crate) mod test_strategies;
